Designed card componenet #3 Refactor - Removed constants folder to import all images - directly importing from the resources folder to the component
