+++ src/components/Icons/BookMarkIcon.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-export const BookMark = () => (
+export const BookMarkIcon = () => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     width="15"
