+++ src/components/Icons/LoveIcon.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-export const Love = () => (
+export const LoveIcon = () => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     width="15"
