+++ src/components/PostsTrendingCard/style.scss
