--- src/constants/Image.js
@@ -1,23 +1,0 @@
-/* global Images */
-
-import momo from "../resources/images/global/momo.png";
-import salad from "../resources/images/global/salad.png";
-import user from "../resources/images/global/user.png";
-import sandwich from "../resources/images/global/sandwich.png";
-import itsnp from "../resources/images/global/itsnp.png";
-/* icons */
-
-import bookmark from "../resources/images/icons/bookmark.png";
-import chat from "../resources/images/icons/chat.png";
-import love from "../resources/images/icons/love.png";
-
-export default {
-  momo,
-  salad,
-  user,
-  itsnp,
-  sandwich,
-  bookmark,
-  chat,
-  love,
-};
