--- src/constants/index.js
@@ -1,3 +1,0 @@
-import AllImage from "./Image";
-
-export { AllImage };
